Added App class and moved main program code into it
